# Conflicts:
#	README.md
#	app/build.gradle
#	app/src/main/java/com/imcys/bilibilias/base/model/user/DownloadTaskBean.kt
#	app/src/main/java/com/imcys/bilibilias/base/utils/DialogUtils.kt
#	app/src/main/java/com/imcys/bilibilias/base/utils/DownloadQueue.kt
#	app/src/main/java/com/imcys/bilibilias/home/ui/activity/AsVideoActivity.kt
#	app/src/main/java/com/imcys/bilibilias/home/ui/activity/tool/MergeVideoActivity.kt
#	app/src/main/java/com/imcys/bilibilias/home/ui/adapter/DownloadFinishTaskAd.kt
#	app/src/main/java/com/imcys/bilibilias/home/ui/fragment/DownloadFragment.kt
#	app/src/main/java/com/imcys/bilibilias/home/ui/fragment/HomeFragment.kt
#	app/src/main/java/com/imcys/bilibilias/home/ui/fragment/ToolFragment.kt
#	app/src/main/java/com/imcys/bilibilias/home/ui/model/view/AsVideoViewModel.kt
#	app/src/main/java/com/imcys/bilibilias/home/ui/model/view/FragmentHomeViewModel.kt
#	app/src/main/java/com/imcys/bilibilias/home/ui/model/view/RCMDVideoModel.kt
#	app/src/main/java/com/imcys/bilibilias/home/ui/viewmodel/AsLoginBsViewModel.kt
#	app/src/main/java/com/imcys/bilibilias/home/ui/viewmodel/MargeVideoViewModel.kt
#	app/src/main/java/com/imcys/bilibilias/home/ui/viewmodel/UserViewModel.kt
#	app/src/main/java/com/imcys/bilibilias/home/ui/viewmodel/factory/AsLoginBsViewModelFactory.kt
#	app/src/main/res/layout/activity_merge_video.xml
#	app/src/main/res/layout/dialog_as_login_bottomsheet.xml
#	app/src/main/res/layout/fragment_download.xml
#	app/src/main/res/layout/item_download_task_finish.xml
#	common/build.gradle
#	common/src/main/java/com/imcys/bilibilias/common/base/utils/AsVideoNumUtils.kt
#	common/src/main/java/com/imcys/bilibilias/common/data/AppDatabase.kt
#	common/src/main/java/com/imcys/bilibilias/common/data/dao/DownloadFinishTaskDao.kt
#	common/src/main/java/com/imcys/bilibilias/common/data/entity/DownloadFinishTaskInfo.kt
#	common/src/main/java/com/imcys/bilibilias/common/data/repository/DownloadFinishTaskRepository.kt

+++ common/src/main/java/com/imcys/bilibilias/common/base/utils/AsVideoNumUtils.kt
@@ -17,7 +17,7 @@ object AsVideoNumUtils {
         if (epRegex.containsMatchIn(string)) return VideoNumConversion.toBvidOffline(
             epRegex.find(
                 string
-            )?.value!!.toInt()
+            )?.value!!.toInt().toLong()
         )
